+++ src/dppd/__init__.py
@@ -5,8 +5,8 @@ try:
     # Change here if project is renamed and does not equal the package name
     dist_name = __name__
     __version__ = get_distribution(dist_name).version
-except DistributionNotFound:  # pragma: no cover
-    __version__ = 'unknown'
+except DistributionNotFound as e:  # pragma: no cover
+    __version__ = 'unknown' + str(e)
 finally:
     del get_distribution, DistributionNotFound