+++ tests/test_single_verbs.py
@@ -38,6 +38,14 @@ def test_head():
     assert_frame_equal(should, actual)
 
 
+def test_ends():
+    df = pd.DataFrame({"a": list(range(10))})
+    actual = dp(df).ends(2).pd
+    should = df.head(2)
+    should = should.append(df.tail(2))
+    assert_frame_equal(should, actual)
+
+
 def test_2_stage_concat():
     df = pd.DataFrame({"a": list(range(10))})
     a = dp(df).head(5).pd
@@ -902,5 +910,11 @@ def test_iter_tuples_in_group_by():
     actual = {k: list(v) for (k, v) in dp(mtcars).groupby("cyl").itertuples()}
     should = {}
     for key, sub_df in mtcars.groupby("cyl"):
-        should[key, ] = list(sub_df.itertuples())
+        should[key,] = list(sub_df.itertuples())
     assert actual == should
+
+
+def test_natsort():
+    df = pd.DataFrame({"a": ["1", "16", "2"], "b": ["another", "category", "level"]})
+    df = dp(df).natsort("a").pd
+    assert (df["a"] == ["1", "2", "16"]).all()