+++ CHANGELOG.rst
@@ -4,6 +4,10 @@ Changelog
 
 
 unreleased
+==========
+
+ - support for itertuples on groupBy objects
+ - column specs now support types (and forward the query to select_dtypes)
 
 0.17
 ==========