+++ CHANGELOG.rst
@@ -8,7 +8,9 @@ unreleased
 
 - support for itertuples on groupBy objects
 - column specs now support types (and forward the query to select_dtypes)
+- column spec now accepts [True] as 'same columns, but sorted alphabetically'
 - ends() for DataFrames
+
 
 0.17
 ==========